dev-requirements.txt
# Package dependancies
Django>=2.2, <4.1
hawk-server==0.0.12
django-hawk==0.0.2

# Testing
tox
freezegun
coverage
factory_boy

# Linting and type checking
isort
flake8
black
mypy
django-stubs
types-freezegun
